# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Instagram Tools API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "default_health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Default Health"
                }
              }
            }
          }
        }
      }
    },
    "/tools/download/formats": {
      "post": {
        "tags": [
          "download"
        ],
        "summary": "Formats",
        "operationId": "download_formats",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormatsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/download/file": {
      "get": {
        "tags": [
          "download"
        ],
        "summary": "File",
        "operationId": "download_file",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "best",
              "title": "Quality"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/download/cover": {
      "post": {
        "tags": [
          "download"
        ],
        "summary": "Cover",
        "operationId": "download_cover",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/download/image": {
      "get": {
        "tags": [
          "download"
        ],
        "summary": "Image",
        "description": "Stream an Instagram CDN image back as a download (covers, profile pics).",
        "operationId": "download_image",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "image.jpg",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/download/zip": {
      "post": {
        "tags": [
          "download"
        ],
        "summary": "Zip Reels",
        "operationId": "download_zip_reels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/profile/reels": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "List Reels",
        "operationId": "profile_list_reels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileReelsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileReelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/profile/info": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Info",
        "operationId": "profile_info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInfoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/transcribe": {
      "post": {
        "tags": [
          "transcribe"
        ],
        "summary": "Start",
        "operationId": "transcribe_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/transcribe/{job_id}": {
      "get": {
        "tags": [
          "transcribe"
        ],
        "summary": "Status",
        "operationId": "transcribe_status",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoverRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Instagram reel/post URL."
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CoverRequest"
      },
      "CoverResponse": {
        "properties": {
          "shortcode": {
            "type": "string",
            "title": "Shortcode"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          }
        },
        "type": "object",
        "required": [
          "shortcode"
        ],
        "title": "CoverResponse"
      },
      "FormatsRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Instagram reel/post URL."
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "FormatsRequest"
      },
      "FormatsResponse": {
        "properties": {
          "shortcode": {
            "type": "string",
            "title": "Shortcode"
          },
          "qualities": {
            "items": {
              "$ref": "#/components/schemas/QualityOption"
            },
            "type": "array",
            "title": "Qualities"
          },
          "audio_available": {
            "type": "boolean",
            "title": "Audio Available"
          }
        },
        "type": "object",
        "required": [
          "shortcode",
          "qualities",
          "audio_available"
        ],
        "title": "FormatsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "error"
            ],
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobResponse"
      },
      "ProfileInfoRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Instagram username (with or without @)."
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "ProfileInfoRequest"
      },
      "ProfileInfoResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name",
            "default": ""
          },
          "biography": {
            "type": "string",
            "title": "Biography",
            "default": ""
          },
          "follower_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follower Count"
          },
          "following_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Following Count"
          },
          "post_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Count"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "default": false
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private",
            "default": false
          },
          "profile_pic_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Pic Url"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "ProfileInfoResponse"
      },
      "ProfileReelsRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Instagram username (with or without @)."
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor",
            "description": "Opaque cursor from a previous page; omit for page 1."
          },
          "page_size": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "Reels per page.",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "ProfileReelsRequest"
      },
      "ProfileReelsResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "reels": {
            "items": {
              "$ref": "#/components/schemas/ReelSummaryModel"
            },
            "type": "array",
            "title": "Reels"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "username",
          "reels"
        ],
        "title": "ProfileReelsResponse"
      },
      "QualityOption": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "filesize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filesize"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "QualityOption"
      },
      "ReelSummaryModel": {
        "properties": {
          "shortcode": {
            "type": "string",
            "title": "Shortcode"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "caption": {
            "type": "string",
            "title": "Caption",
            "default": ""
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags",
            "default": []
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "type": "object",
        "required": [
          "shortcode",
          "url"
        ],
        "title": "ReelSummaryModel"
      },
      "Segment": {
        "properties": {
          "start": {
            "type": "number",
            "title": "Start"
          },
          "end": {
            "type": "number",
            "title": "End"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "text"
        ],
        "title": "Segment"
      },
      "TranscribeRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Instagram reel/post URL."
          },
          "engine": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local_whisper",
                  "openai",
                  "assemblyai"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine",
            "description": "Override the default engine."
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "TranscribeRequest"
      },
      "TranscriptResult": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Segment"
            },
            "type": "array",
            "title": "Segments",
            "default": []
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "caption": {
            "type": "string",
            "title": "Caption",
            "default": ""
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TranscriptResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ZipRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Urls"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "description": "Quality id, or \"audio\".",
            "default": "best"
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "ZipRequest"
      }
    }
  }
}